+++ test/integration/_common.js
@@ -22,11 +22,17 @@ export function getTestFn(shell) {
     return test.skip;
   }
 
-  try {
-    if (!isCI && typeof shell === "string") {
-      which.sync(shell, { path: process.env.PATH || process.env.Path });
-    }
+  if (isCI) {
+    return test;
+  }
 
+  if (typeof shell !== "string") {
+    return test;
+  }
+
+  const PATH = process.env.PATH || process.env.Path;
+  try {
+    which.sync(shell, { path: PATH });
     return test;
   } catch {
     return test.skip;