+++ CHANGELOG.md
@@ -9,6 +9,17 @@ Versioning].
 
 ## Unreleased
 
+- _No changes yet_
+
+## 3.0.0 (2026-06-13)
+
+- BREAKING CHANGE: Drop support for `which` versions `^3`, `^4`, `^5`, and
+  `^6`. ([#2572])
+- BREAKING CHANGE: Drop support for Node.js `^14.18.0`, `^16.13.0`, `18.0.0`
+  through `22.22.2`, `24.0.0` through `24.14.0`, and `^25`. ([#2572])
+- BREAKING CHANGE: Limit CommonJS support to loading ESModules using
+  `require()`. ([#2572])
+- BREAKING CHANGE: Remove `Shescape` export from the testing module. ([#2572])
 - Add support for `which` v7. ([#2534])
 
 ## 2.1.12 (2026-05-06)
@@ -417,6 +428,7 @@ Versioning].
 [#2490]: https://github.com/ericcornelissen/shescape/pull/2490
 [#2520]: https://github.com/ericcornelissen/shescape/pull/2520
 [#2534]: https://github.com/ericcornelissen/shescape/pull/2534
+[#2572]: https://github.com/ericcornelissen/shescape/pull/2572
 [552e8ea]: https://github.com/ericcornelissen/shescape/commit/552e8eab56861720b1d4e5474fb65741643358f9
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/
 [semantic versioning]: https://semver.org/spec/v2.0.0.html